Address review feedback: rename c->config, add emulation warning, fix mypy
- Rename `c` parameter to `config` in NvFp4LinearKernel base class and all
  subclasses for clarity (fxmarty-amd review)
- Restore emulation fallback warning in init_nvfp4_linear_kernel() that lists
  reasons other backends are unavailable (fxmarty-amd review)
- Rename NvFp4LinearKernel.py -> base.py to fix mypy module/class name collision
