+++ vllm/model_executor/kernels/linear/nvfp4/marlin.py
@@ -10,7 +10,7 @@ from vllm.model_executor.layers.quantization.utils.marlin_utils_fp4 import (
     prepare_fp4_layer_for_marlin,
 )
 
-from .NvFp4LinearKernel import NvFp4LinearKernel, NvFp4LinearLayerConfig
+from .base import NvFp4LinearKernel, NvFp4LinearLayerConfig
 
 logger = init_logger(__name__)
 
@@ -27,7 +27,7 @@ class MarlinNvFp4LinearKernel(NvFp4LinearKernel):
         return False, "Marlin FP4 not available"
 
     @classmethod
-    def can_implement(cls, c: NvFp4LinearLayerConfig) -> tuple[bool, str | None]:
+    def can_implement(cls, config: NvFp4LinearLayerConfig) -> tuple[bool, str | None]:
         return True, None
 
     def process_weights_after_loading(self, layer: torch.nn.Module) -> None:
